+++ tencentcloud/cbs/v20170312/models.py
@@ -1517,6 +1517,8 @@ class Disk(AbstractModel):
         :type Shareable: bool
         :param CreateTime: 云硬盘的创建时间。
         :type CreateTime: str
+        :param DeleteSnapshot: 销毁云盘时删除关联的非永久保留快照。0 表示非永久快照不随云盘销毁而销毁，1表示非永久快照随云盘销毁而销毁，默认取0。快照是否永久保留可以通过DescribeSnapshots接口返回的快照详情的IsPermanent字段来判断，true表示永久快照，false表示非永久快照。
+        :type DeleteSnapshot: int
         """
         self.DeleteWithInstance = None
         self.RenewFlag = None
@@ -1553,6 +1555,7 @@ class Disk(AbstractModel):
         self.ReturnFailCode = None
         self.Shareable = None
         self.CreateTime = None
+        self.DeleteSnapshot = None
 
 
     def _deserialize(self, params):
@@ -1598,6 +1601,7 @@ class Disk(AbstractModel):
         self.ReturnFailCode = params.get("ReturnFailCode")
         self.Shareable = params.get("Shareable")
         self.CreateTime = params.get("CreateTime")
+        self.DeleteSnapshot = params.get("DeleteSnapshot")
         memeber_set = set(params.keys())
         for name, value in vars(self).items():
             if name in memeber_set: