+++ tencentcloud/ivld/v20210903/errorcodes.py
@@ -14,6 +14,21 @@
 # limitations under the License.
 
 
+# SecredId失效。
+AUTHFAILURE_INVALIDSECRETID = 'AuthFailure.InvalidSecretId'
+
+# MFA失败。
+AUTHFAILURE_MFAFAILURE = 'AuthFailure.MFAFailure'
+
+# SecretId不存在。
+AUTHFAILURE_SECRETIDNOTFOUND = 'AuthFailure.SecretIdNotFound'
+
+# 签名已过期。
+AUTHFAILURE_SIGNATUREEXPIRE = 'AuthFailure.SignatureExpire'
+
+# 签名校验失败。
+AUTHFAILURE_SIGNATUREFAILURE = 'AuthFailure.SignatureFailure'
+
 # 任务已完成。
 AUTHFAILURE_TASKFINISHED = 'AuthFailure.TaskFinished'
 
@@ -31,6 +46,9 @@ FAILEDOPERATION_AITEMPLATENOTEXIST = 'FailedOperation.AiTemplateNotExist'
 
 # 自定义人物分类已存在。
 FAILEDOPERATION_CATEGORYEXIST = 'FailedOperation.CategoryExist'
+
+# 自定义类型层级变化。
+FAILEDOPERATION_CATEGORYLEVELCHANGED = 'FailedOperation.CategoryLevelChanged'
 
 # 自定义人物分类被引用，不能删除。
 FAILEDOPERATION_CATEGORYREFERRED = 'FailedOperation.CategoryReferred'
@@ -134,10 +152,10 @@ INVALIDPARAMETER_INVALIDFILEPATH = 'InvalidParameter.InvalidFilePath'
 # 图片ID不合法。
 INVALIDPARAMETER_INVALIDIMAGEID = 'InvalidParameter.InvalidImageId'
 
-# 一级自定义标签不合法。
+# 一级自定义类型不合法。
 INVALIDPARAMETER_INVALIDL1CATEGORY = 'InvalidParameter.InvalidL1Category'
 
-# 二级自定义标签不合法。
+# 二级自定义类型不合法。
 INVALIDPARAMETER_INVALIDL2CATEGORY = 'InvalidParameter.InvalidL2Category'
 
 # MD5不合法。
@@ -211,6 +229,12 @@ REQUESTLIMITEXCEEDED_BATCHIMPORTOVERFLOW = 'RequestLimitExceeded.BatchImportOver
 
 # 同时发起过多任务。
 REQUESTLIMITEXCEEDED_CONCURRENCYOVERFLOW = 'RequestLimitExceeded.ConcurrencyOverflow'
+
+# 自定义人物类型不存在。
+RESOURCENOTFOUND_CUSTOMCATEGORYNOTFOUND = 'ResourceNotFound.CustomCategoryNotFound'
+
+# 自定义人物库不存在。
+RESOURCENOTFOUND_CUSTOMGROUPNOTFOUND = 'ResourceNotFound.CustomGroupNotFound'
 
 # 媒资文件不存在。
 RESOURCENOTFOUND_MEDIANOTFOUND = 'ResourceNotFound.MediaNotFound'