+++ tencentcloud/sms/v20190711/models.py
@@ -57,12 +57,13 @@ class AddSmsSignRequest(AbstractModel):
 注：不能重复申请已通过或待审核的签名。
         :type SignName: str
         :param SignType: 签名类型。其中每种类型后面标注了其可选的 DocumentType（证明类型）：
-0：公司（0，1，2，3）。
-1：APP（0，1，2，3，4） 。
-2：网站（0，1，2，3，5）。
-3：公众号或者小程序（0，1，2，3，6）。
-4：商标（7）。
-5：政府/机关事业单位/其他机构（2，3）。
+0：公司，可选 DocumentType 有（0，1）。
+1：APP，可选 DocumentType 有（0，1，2，3，4） 。
+2：网站，可选 DocumentType 有（0，1，2，3，5）。
+3：公众号，可选 DocumentType 有（0，1，2，3，8）。
+4：商标，可选 DocumentType 有（7）。
+5：政府/机关事业单位/其他机构，可选 DocumentType 有（2，3）。
+6：小程序，可选 DocumentType 有（0，1，2，3，6）。
 注：必须按照对应关系选择证明类型，否则会审核失败。
         :type SignType: int
         :param DocumentType: 证明类型：
@@ -74,6 +75,7 @@ class AddSmsSignRequest(AbstractModel):
 5：网站备案后台截图（个人开发网站）。
 6：小程序设置页面截图（个人认证小程序）。
 7：商标注册书。
+8：公众号设置页面截图（个人认证公众号）。
         :type DocumentType: int
         :param International: 是否国际/港澳台短信：
 0：表示国内短信。
@@ -754,12 +756,13 @@ class ModifySmsSignRequest(AbstractModel):
         :param SignName: 签名名称。
         :type SignName: str
         :param SignType: 签名类型。其中每种类型后面标注了其可选的 DocumentType（证明类型）：
-0：公司（0，1，2，3）。
-1：APP（0，1，2，3，4） 。
-2：网站（0，1，2，3，5）。
-3：公众号或者小程序（0，1，2，3，6）。
-4：商标（7）。
-5：政府/机关事业单位/其他机构（2，3）。
+0：公司，可选 DocumentType 有（0，1）。
+1：APP，可选 DocumentType 有（0，1，2，3，4） 。
+2：网站，可选 DocumentType 有（0，1，2，3，5）。
+3：公众号，可选 DocumentType 有（0，1，2，3，8）。
+4：商标，可选 DocumentType 有（7）。
+5：政府/机关事业单位/其他机构，可选 DocumentType 有（2，3）。
+6：小程序，可选 DocumentType 有（0，1，2，3，6）。
 注：必须按照对应关系选择证明类型，否则会审核失败。
         :type SignType: int
         :param DocumentType: 证明类型：
@@ -767,10 +770,11 @@ class ModifySmsSignRequest(AbstractModel):
 1：企业营业执照。
 2：组织机构代码证书。
 3：社会信用代码证书。
-4：应用后台管理截图(个人开发APP)。
-5：网站备案后台截图(个人开发网站)。
-6：小程序设置页面截图(个人认证小程序)。
+4：应用后台管理截图（个人开发APP）。
+5：网站备案后台截图（个人开发网站）。
+6：小程序设置页面截图（个人认证小程序）。
 7：商标注册书。
+8：公众号设置页面截图（个人认证公众号）。
         :type DocumentType: int
         :param International: 是否国际/港澳台短信：
 0：表示国内短信。